Add support for ShenZen HOMA controllers
There are two models: one line voltage dimmer (brightness) and a
configurable 24V four-channel PWM controller. Depending on the DIP
switches on the RGBW controller, it appears as a single brightness,
color temperature, or extended color device; or as a combination color
and brightness device (RGBW).

They are sold in Germany through Amazon:
* https://www.amazon.de/gp/product/B076KPQWFW 230V
* https://www.amazon.de/gp/product/B076K6VJ8G 24V RGBW