Add `anyOn` config.json key.
See #609.  When `anyOn` is true (default), homebridge-hue exposes a custom _AnyOn_ characteristic for groups, mapped to `state.any_on`.  In that case, the _On_ characteristic is mapped to `state.all_on`.  Set `anyOn` to false to prevent _AnyOn_ from being exposed.  In this case, _On_ is mapped to `state.any_on`.