Update HueClient.js
Bug fix commands waiting to be executed not executed when API error occurs.

Especially noticeable for lights with `noWaitTimeUpdate` (GLEDOPTO GL-C-008), or when `waitTimeUpdate` has been set to 0.  When turning the light on from Home, it sends `bri` and `on` (unfortunately in that order); the `bri` fails because the light is not yet on, and the `on` would get lost.